src/ingestion/requirements.txt
snowflake-connector-python
python-dotenv
